+++ eventoLocation-service/src/main/java/org/unical/enterprise/eventoLocation/controller/PreferitiController.java
@@ -1,7 +1,10 @@
 package org.unical.enterprise.eventoLocation.controller;
 
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import org.unical.enterprise.eventoLocation.data.dto.PreferitiDto;
+import org.unical.enterprise.eventoLocation.service.EventoService;
 import org.unical.enterprise.eventoLocation.service.PreferitiService;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -13,24 +16,49 @@ import org.springframework.http.ResponseEntity;
 public class PreferitiController {
 
     private final PreferitiService preferitiService;
+    private final EventoService eventoService;
 
     @GetMapping(value="/{id}")
     private ResponseEntity<?> findById(@PathVariable("id") Long id){
         return new ResponseEntity<>(preferitiService.getById(id), HttpStatus.OK);
     }
 
-    ///TODO ERRORI CREAZIONE PREFERITI (duplicated keys)
     @PostMapping(consumes = "application/json")
     @ResponseStatus(HttpStatus.CREATED)
     public ResponseEntity<?> createPreferito(@RequestBody PreferitiDto preferiti) {
-        preferitiService.save(preferiti);
-        return new ResponseEntity<>(HttpStatus.CREATED);
+        if (eventoService.getByIdNoLocation(preferiti.getEvento_id()) != null) {
+            String user= SecurityContextHolder.getContext().getAuthentication().getName();
+            String utenteInserito= preferiti.getUtente_id();
+            if (user.equals(utenteInserito)) {
+                if (preferitiService.getById(preferiti.getId()) == null) {
+                    preferitiService.save(preferiti);
+                    return new ResponseEntity<>(HttpStatus.CREATED);
+                } else {
+                    return new ResponseEntity<>(HttpStatus.CONFLICT);
+                }
+            } else {
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            }
+        } else {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
 
     @DeleteMapping(path="/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deletePreferito(@PathVariable("id") Long id) {
-        preferitiService.delete(id);
+        if (preferitiService.getById(id) != null) {
+            String user= SecurityContextHolder.getContext().getAuthentication().getName();
+            String utenteInserito=preferitiService.getById(id).getUtente_id();
+            if (user.equals(utenteInserito)) {
+                preferitiService.delete(id);
+            } else {
+                throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+            }
+        } else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+
     }
